Final Final Final

+++ application/controllers/deleteContact.php
@@ -31,8 +31,30 @@ class deleteContact extends REST_Controller {
 	{
         
         $id = $this->input->post('contact_id');
+        $type=$this->input->post('data_store');
+        if($type == "DATABASE")
+        {
         $this->db->delete('contacts', array('id'=>$id));
         $this->response(['Item deleted successfully.'], REST_Controller::HTTP_OK);
+        }
+
+        else if($type == "CRM")
+        {
+            $ch = curl_init();
+            $url="https://domain.freshsales.io/api/contacts/$id";
+            $access="TRcytqeHoqGV-49FGcKn6Q";
+            $header[] = 'Content-Type:  application/json';
+            $header[] = 'Authorization: Token token='.$access;
+            curl_setopt($ch, CURLOPT_URL, $url);
+            curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "DELETE");
+           // curl_setopt($ch, CURLOPT_POSTFIELDS, $id);
+            curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+            $result = curl_exec($ch);
+           // $result = json_decode($result);
+            curl_close($ch);
+
+            return $result;
+        }
 	}
